Fixes segmentation contourer from outputting ugly contours
The cause was an error in upscaling.
Found another issue as well when analysing the model api implementation
of the upscaling.
There was a case where the mask would touch the border of the image and
that would cause artificats.
By adding an empty border the contourer is able to fix that.